# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: RipTide API - Comprehensive Specification
  version: 1.0.0
  x-last-validated: "2025-10-10T07:44:00Z"
  description: '# RipTide Crawler API


    A high-performance web scraping and content extraction API with 59 endpoints across
    12 categories.


    ## Phase 1: Core Crawling & Event System (11 endpoints)

    - `/crawl` - Batch URL crawling with adaptive gate system

    - `/crawl/stream`, `/crawl/sse`, `/crawl/ws` - Real-time streaming

    - `/deepsearch`, `/deepsearch/stream` - Web search integration

    - `/render` - Headless browser rendering

    - `/healthz`, `/metrics` - System health and Prometheus metrics


    ## Phase 2: Advanced Extraction (14 endpoints)

    - `/strategies/*` - Multi-strategy extraction (CSS, WASM, LLM, Regex)

    - `/pdf/*` - PDF processing with progress streaming

    - `/api/v1/tables/*` - Table extraction and export (CSV/Markdown)


    ## Phase 3: Enterprise Features (34 endpoints)

    - `/spider/*` - Deep crawling with frontier management (3)

    - `/stealth/*` - Stealth configuration and testing (4)

    - `/api/v1/llm/*` - LLM provider management (4)

    - `/sessions/*` - Session and cookie management (12)

    - `/workers/*` - Async job queue and scheduling (9)

    - `/monitoring/*` - Health monitoring and alerts (6)


    ## Architecture Highlights

    - **Dual-Path Pipeline**: Fast CSS + async AI enhancement

    - **Event-Driven**: Core event bus for monitoring

...

...


...

    '
  contact:
...
servers:
...
paths:
...
  /api/v1/tables/{id}/export:
    get:
      tags:
      - Tables
      summary: Export table (CSV/Markdown)
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Table identifier
      responses:
        '200':
          description: Success
      operationId: api_v1_tables_id_export_get
...